% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caseNamesKspm.R
\name{case.names.kspm}
\alias{case.names.kspm}
\title{Case names of fitted models}
\usage{
\method{case.names}{kspm}(object, ...)
}
\arguments{
\item{object}{an object of class "kspm", usually, a result of a call to \code{kspm}.}

\item{...}{additional optional argument (currently unused).}
}
\value{
a character vector.
}
\description{
Simple utility returning names of cases involved in a kernel semi parametric model.
}
\seealso{
\link{kspm} for fitting model, \link{nobs.kspm}, \link{variable.names.kspm}.
}
\author{
Catherine Schramm, Aurelie Labbe, Celia Greenwood
}
