% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy.R
\docType{data}
\name{energy}
\alias{energy}
\title{Energy consumption measuring hourly during 22 days}
\format{A data frame with 504 rows and 7 variables:
\describe{
  \item{power}{energy consumption}
  \item{date}{date}
  \item{T}{temperature}
  \item{P}{pression}
  \item{HR}{humidity rate}
  \item{hour}{hour (categorical)}
  \item{hour.num}{hour (numerical)}
}}
\source{
\url{https://iles-ponant-edf-sei.opendatasoft.com}, \url{www.infoclimat.fr}
}
\usage{
energy
}
\description{
A dataset containing the energy consumption and other attributes during 22 days.
}
\keyword{datasets}
