% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParameters.R
\name{get.parameters}
\alias{get.parameters}
\title{compute Kernel Semi Parametric model parameters}
\usage{
get.parameters(X = NULL, Y = NULL, kernelList = NULL,
  free.parameters = NULL, n = NULL, not.missing = NULL,
  compute.kernel = NULL)
}
\arguments{
\item{X}{X matrix}

\item{Y}{response matrix}

\item{kernelList}{list of kernels}

\item{free.parameters}{free parameters}

\item{n}{number of samples}

\item{not.missing}{number of non missing samples}

\item{compute.kernel}{boolean indicating if kernel should be computed}
}
\description{
internal function to compute model parameters
}
\author{
Catherine Schramm, Aurelie Labbe, Celia Greenwood
}
