% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infoKspm.R
\name{info.kspm}
\alias{info.kspm}
\title{Giving information about Kernel Semi parametric Model Fits}
\usage{
info.kspm(object, print = TRUE)
}
\arguments{
\item{object}{an object of class "kspm", usually, a result of a call to \code{kspm}.}

\item{print}{logical, if \code{TRUE}, table of information are printed.}
}
\value{
\code{info.kspm} returns a table of information whose each row corresponds to a kernel included in the model and columns are:
  \item{type}{type of object used to define the kernel}
  \item{dim}{dimension of data used in the model}
  \item{type.predict}{type of object the user should provide in \link{predict.kspm} function}
  \item{dim.predict}{dimension of object the user should provide in \link{predict.kspm} function}
}
\description{
gives information about Kernel Semi parametric Model Fits
}
\seealso{
\link{kspm}, \link{predict.kspm}
}
\author{
Catherine Schramm, Aurelie Labbe, Celia Greenwood
}
