% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernelList.R
\name{kernel.list}
\alias{kernel.list}
\title{List of kernel parts included in the kernel semi parametric model}
\usage{
kernel.list(formula, data, names)
}
\arguments{
\item{formula}{kernel part formula provided in the \code{kspm} function.}

\item{data}{data provided in the \code{kspm} function.}

\item{names}{row names of samples as they are evaluated in \code{kspm} function.}
}
\description{
internal method for listing all kernel parts included in the model
}
\author{
Catherine Schramm, Aurelie Labbe, Celia Greenwood
}
