% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernelMethods.R
\name{kernel.method}
\alias{kernel.method}
\alias{comb}
\alias{check.integer}
\alias{asOneSidedFormula}
\alias{splitFormula}
\alias{computes.Kernel}
\alias{computes.Kernel.interaction}
\alias{computes.KernelALL}
\alias{renames.Kernel}
\alias{objects.Kernel}
\title{some internal methods in computation of kernel semi parametric model}
\usage{
comb(x, ...)

check.integer(N)

asOneSidedFormula(object)

splitFormula(form, sep = "/")

computes.Kernel(x, ind, nameKernel, not.missing = NULL)

computes.Kernel.interaction(x, ind, nameKernel, not.missing = NULL)

computes.KernelALL(kernelList, not.missing = NULL)

renames.Kernel(object, names)

objects.Kernel(formula)
}
\arguments{
\item{x}{list of objects}

\item{...}{other arguments}

\item{N}{numeric value}

\item{object}{formula provided in the kernel part of \code{kspm} function}

\item{form}{formula}

\item{sep}{separator}

\item{ind}{index value}

\item{nameKernel}{name of kernel}

\item{not.missing}{non missing values}

\item{kernelList}{list of kernels}

\item{names}{name of kernel}

\item{formula}{formula}
}
\description{
internal methods
}
\author{
Catherine Schramm, Aurelie Labbe, Celia Greenwood
}
