% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lossFunction.R
\name{lossFunction.looe}
\alias{lossFunction.looe}
\title{Computation of the leave one out error (LOOE) in kernel semi parametric model}
\usage{
lossFunction.looe(param. = NULL, Y. = NULL, X. = NULL,
  kernelList. = NULL, n. = NULL, not.missing. = NULL,
  compute.kernel. = NULL, print.lambda. = FALSE)
}
\arguments{
\item{param.}{initial parameter values.}

\item{Y.}{response matrix.}

\item{X.}{X matrix (linear part).}

\item{kernelList.}{list of kernels (kernel part).}

\item{n.}{nb of samples.}

\item{not.missing.}{nb of non missing samples.}

\item{compute.kernel.}{boolean. If TRUE, the kernel matrix is computed at each iteration. Should be TRUE when hyperparameters of kernel functions should be estimated by the model.}

\item{print.lambda.}{boolean. If TRUE, values of tunning parameters (lambda) are printed at each iteration.}
}
\description{
internal function to optimize model for estimating hyperparameters based on LOOE
}
\author{
Catherine Schramm, Aurelie Labbe, Celia Greenwood
}
