% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printKspm.R
\name{print.kspm}
\alias{print.kspm}
\alias{print.summary.kspm}
\title{Print results from a Kernel Semi parametric Model Fit}
\usage{
\method{print}{kspm}(x, ...)

\method{print}{summary.kspm}(x, ...)
}
\arguments{
\item{x}{an object used to select a method. Usually, a result of a call to \code{kspm} or a result from \code{summary.kspm}.}

\item{...}{additional optional argument (currently unused).}
}
\description{
print method for class "kspm".
}
\seealso{
\link{kspm} for fitting model, \link{summary.kspm}
}
\author{
Catherine Schramm, Aurelie Labbe, Celia Greenwood
}
