% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchParameters.R
\name{search.parameters}
\alias{search.parameters}
\title{Optimisation to cumpute hyperparameter in Kernel Semi Parametric model}
\usage{
search.parameters(Y = NULL, X = NULL, kernelList = NULL, n = NULL,
  not.missing = NULL, compute.kernel = NULL, controlKspm = NULL)
}
\arguments{
\item{Y}{response matrix}

\item{X}{X matrix}

\item{kernelList}{of kernels}

\item{n}{nb of samples}

\item{not.missing}{nb of non missing samples}

\item{compute.kernel}{boolean kernel computation}

\item{controlKspm}{control parameters}
}
\description{
internal function to optimize model for estimating hyperparameters
}
\author{
Catherine Schramm, Aurelie Labbe, Celia Greenwood
}
