% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigmaKspm.R
\name{sigma.kspm}
\alias{sigma.kspm}
\title{Extract residuals standard deviation}
\usage{
\method{sigma}{kspm}(object, ...)
}
\arguments{
\item{object}{an object of class "kspm", usually, a result of a call to \code{kspm}.}

\item{...}{additional optional argument (currently unused).}
}
\value{
typically a number, the estimated standard deviation of the errors ("residual standard deviation")
}
\description{
Returns the residuals standard deviation (sigma) for object of class "kspm".
}
\details{
The value returned by the method is \eqn{\sqrt{\frac{RSS}{edf}}}{sqrt(RSS / edf)} where \eqn{RSS}{RSS} is the residual sum of squares and \eqn{edf}{edf} is the effective degree of freedom.
}
\seealso{
\link{kspm} for fitting model, \link{summary.kspm}, \link{residuals.kspm}, \link{nobs.kspm}, \link{deviance.kspm}.
}
\author{
Catherine Schramm, Aurelie Labbe, Celia Greenwood
}
