% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernelTest.R
\name{test.function}
\alias{test.function}
\alias{test.1.kernel}
\alias{test.global.kernel}
\alias{test.k.kernel}
\title{Score Tests for kernel part in kernel semi parametric model}
\usage{
test.1.kernel(object)

test.global.kernel(object)

test.k.kernel(object, kernel.name)
}
\arguments{
\item{object}{an object of class "kspm"}

\item{kernel.name}{vector of character listing names of kernels for which test should be performed}
}
\value{
p values
}
\description{
Perform score tests for kernel part in kernel semi parametric model
}
\references{
Schweiger, Regev, et al. "RL SKAT: an exact and efficient score test for heritability and set tests." Genetics (2017): genetics 300395.

Li, Shaoyu, and Yuehua Cui. "Gene centric gene gene interaction: A model based kernel machine method." The Annals of Applied Statistics 6.3 (2012): 1134:1161.

Oualkacha, Karim, et al. "Adjusted sequence kernel association test for rare variants controlling for cryptic and family relatedness." Genetic epidemiology 37.4 (2013): 366:376.

Ge, Tian, et al. "A kernel machine method for detecting effects of interaction between multidimensional variable sets: An imaging genetics application." Neuroimage 109 (2015): 505:514.
}
\author{
Catherine Schramm, Aurelie Labbe, Celia Greenwood
}
